+++ fe/src/components/issueDetail/detailContent/index.jsx
@@ -1,6 +1,7 @@
 import { CommentBody, CommentBox, CommentHeaderBox, TimeStampBox } from './style';
 import UserImageLarge from '@assets/images/UserImageLarge.svg';
-import Label from '@components/common/lable';
+import Label from '@components/common/label';
+import TextArea from '@components/common/textArea';
 
 const CommentHeader = ({ image, assignee, timeStamp }) => {
   return (
@@ -15,10 +16,13 @@ const CommentHeader = ({ image, assignee, timeStamp }) => {
 
 const DetailContent = () => {
   return (
-    <CommentBox>
-      <CommentHeader image={UserImageLarge} assignee="daon" timeStamp="9분" />
-      <CommentBody>이슈 트래커 언제 끝나아</CommentBody>
-    </CommentBox>
+    <>
+      <CommentBox>
+        <CommentHeader image={UserImageLarge} assignee="daon" timeStamp="9분" />
+        <CommentBody>이슈 트래커 언제 끝나아</CommentBody>
+      </CommentBox>
+      <TextArea />
+    </>
   );
 };
